Now th125 works with native d3dx9 and with WINEDIB off.  WINEDIB=ON is broken for now.

Conflicts:
	configure
	dlls/gdi32/driver.c

+++ dlls/msi/tests/format.c
@@ -201,15 +201,20 @@ static MSIHANDLE create_package_db(void)
     DeleteFile(msifile);
 
     /* create an empty database */
-    res = MsiOpenDatabase(msifile, MSIDBOPEN_CREATE, &hdb );
-    ok( res == ERROR_SUCCESS , "Failed to create database\n" );
+    res = MsiOpenDatabase(msifile, MSIDBOPEN_CREATEDIRECT, &hdb );
+    ok( res == ERROR_SUCCESS , "Failed to create database %u\n", res );
     if( res != ERROR_SUCCESS )
-        return hdb;
+        return 0;
 
     res = MsiDatabaseCommit( hdb );
     ok( res == ERROR_SUCCESS , "Failed to commit database\n" );
+    if( res != ERROR_SUCCESS )
+        return 0;
 
     res = set_summary_info(hdb);
+    ok( res == ERROR_SUCCESS , "Failed to set summary info %u\n", res );
+    if( res != ERROR_SUCCESS )
+        return 0;
 
     res = run_query( hdb,
             "CREATE TABLE `Directory` ( "
@@ -217,25 +222,31 @@ static MSIHANDLE create_package_db(void)
             "`Directory_Parent` CHAR(255), "
             "`DefaultDir` CHAR(255) NOT NULL "
             "PRIMARY KEY `Directory`)" );
-    ok( res == ERROR_SUCCESS , "Failed to create directory table\n" );
+    ok( res == ERROR_SUCCESS , "Failed to create directory table %u\n", res );
 
     return hdb;
 }
 
-static MSIHANDLE package_from_db(MSIHANDLE hdb)
+static UINT package_from_db(MSIHANDLE hdb, MSIHANDLE *handle)
 {
     UINT res;
-    CHAR szPackage[10];
+    CHAR szPackage[12];
     MSIHANDLE hPackage;
 
-    sprintf(szPackage,"#%i",hdb);
-    res = MsiOpenPackage(szPackage,&hPackage);
-    ok( res == ERROR_SUCCESS , "Failed to open package\n" );
+    sprintf(szPackage, "#%u", hdb);
+    res = MsiOpenPackage(szPackage, &hPackage);
+    if (res != ERROR_SUCCESS)
+        return res;
 
     res = MsiCloseHandle(hdb);
-    ok( res == ERROR_SUCCESS , "Failed to close db handle\n" );
+    if (res != ERROR_SUCCESS)
+    {
+        MsiCloseHandle(hPackage);
+        return res;
+    }
 
-    return hPackage;
+    *handle = hPackage;
+    return ERROR_SUCCESS;
 }
 
 static void create_test_file(const CHAR *name)
@@ -250,22 +261,21 @@ static void create_test_file(const CHAR *name)
     CloseHandle(file);
 }
 
-static MSIHANDLE helper_createpackage( const char *szName )
+static UINT helper_createpackage( const char *szName, MSIHANDLE *handle )
 {
-    MSIHANDLE hdb = 0;
+    MSIHANDLE hPackage, suminfo, hdb = 0;
     UINT res;
-    CHAR szPackage[10];
-    MSIHANDLE hPackage;
-    MSIHANDLE suminfo;
 
     DeleteFile(szName);
 
     /* create an empty database */
-    res = MsiOpenDatabase(szName, MSIDBOPEN_CREATE, &hdb );
-    ok( res == ERROR_SUCCESS , "Failed to create database\n" );
+    res = MsiOpenDatabase(szName, MSIDBOPEN_CREATEDIRECT, &hdb );
+    ok( res == ERROR_SUCCESS , "Failed to create database %u\n", res );
+    if (res != ERROR_SUCCESS)
+        return res;
 
     res = MsiDatabaseCommit( hdb );
-    ok( res == ERROR_SUCCESS , "Failed to commit database\n" );
+    ok( res == ERROR_SUCCESS , "Failed to commit database %u\n", res );
 
     /* build summary info */
     res = MsiGetSummaryInformation(hdb, NULL, 7, &suminfo);
@@ -303,14 +313,15 @@ static MSIHANDLE helper_createpackage( const char *szName )
     res = MsiCloseHandle( suminfo);
     ok( res == ERROR_SUCCESS , "Failed to close suminfo\n" );
 
-    sprintf(szPackage,"#%i",hdb);
-    res = MsiOpenPackage(szPackage,&hPackage);
-    ok( res == ERROR_SUCCESS , "Failed to open package\n" );
+    res = package_from_db( hdb, &hPackage );
+    MsiCloseHandle(hdb);
 
-    res = MsiCloseHandle( hdb );
-    ok( res == ERROR_SUCCESS , "Failed to close database\n" );
+    if (res != ERROR_SUCCESS)
+        DeleteFileA( szName );
+    else
+        *handle = hPackage;
 
-    return hPackage;
+    return res;
 }
 
 static void test_createpackage(void)
@@ -318,11 +329,16 @@ static void test_createpackage(void)
     MSIHANDLE hPackage = 0;
     UINT res;
 
-    hPackage = helper_createpackage( msifile );
-    ok ( hPackage != 0, " Failed to create package\n");
+    res = helper_createpackage( msifile, &hPackage );
+    if (res == ERROR_INSTALL_PACKAGE_REJECTED)
+    {
+        skip("Not enough rights to perform tests\n");
+        return;
+    }
+    ok( res == ERROR_SUCCESS, "Failed to create package %u\n", res );
 
-    res = MsiCloseHandle( hPackage);
-    ok( res == ERROR_SUCCESS , "Failed to close package\n" );
+    res = MsiCloseHandle( hPackage );
+    ok( res == ERROR_SUCCESS , "Failed to close package %u\n", res );
 
     DeleteFile( msifile );
 }
@@ -1625,8 +1641,13 @@ static void test_formatrecord_package(void)
     UINT r;
     DWORD sz=100;
 
-    package = helper_createpackage( msifile );
-    ok(package!=0, "Unable to create package\n");
+    r = helper_createpackage( msifile, &package );
+    if (r == ERROR_INSTALL_PACKAGE_REJECTED)
+    {
+        skip("Not enough rights to perform tests\n");
+        return;
+    }
+    ok( r == ERROR_SUCCESS, "Unable to create package %u\n", r );
 
     hrec = MsiCreateRecord(12);
     ok( hrec, "failed to create record\n");
@@ -2120,7 +2141,7 @@ static void test_formatrecord_package(void)
 
 static void test_formatrecord_tables(void)
 {
-    MSIHANDLE hdb, hpkg, hrec;
+    MSIHANDLE hdb, hrec, hpkg = 0;
     CHAR buf[MAX_PATH];
     CHAR curr_dir[MAX_PATH];
     CHAR expected[MAX_PATH];
@@ -2197,8 +2218,15 @@ static void test_formatrecord_tables(void)
     r = add_custom_action_entry( hdb, "'EscapeIt3', 51, 'prop', '[abcd\\xefgh]'" );
     ok( r == ERROR_SUCCESS, "cannt add custom action: %d\n", r);
 
-    hpkg = package_from_db( hdb );
-    ok( hpkg, "failed to create package\n");
+    r = package_from_db( hdb, &hpkg );
+    if (r == ERROR_INSTALL_PACKAGE_REJECTED)
+    {
+        skip("Not enough rights to perform tests\n");
+        MsiCloseHandle( hdb );
+        DeleteFile( msifile );
+        return;
+    }
+    ok( r == ERROR_SUCCESS, "failed to create package %u\n", r );
 
     MsiCloseHandle( hdb );
 
@@ -2381,12 +2409,16 @@ static void test_formatrecord_tables(void)
 
 static void test_processmessage(void)
 {
-    MSIHANDLE hrec;
-    MSIHANDLE package;
-    int r;
+    MSIHANDLE hrec, package;
+    UINT r;
 
-    package = helper_createpackage( msifile );
-    ok(package!=0, "Unable to create package\n");
+    r = helper_createpackage( msifile, &package );
+    if (r == ERROR_INSTALL_PACKAGE_REJECTED)
+    {
+        skip("Not enough rights to perform tests\n");
+        return;
+    }
+    ok( r == ERROR_SUCCESS, "Unable to create package %u\n", r );
 
     hrec = MsiCreateRecord(3);
     ok( hrec, "failed to create record\n");
